fix(nvme-ptpl): don't error out when there's no holder
When loading the reservation file, the restore was erroring out when there was no holder.
This is not an error case, as it might be that no one was the holder prior to power loss.
Check if there was any holder and if not, simply leave the namespace in similar state.

Signed-off-by: Tiago Castro <tiagolobocastro@gmail.com>
Signed-off-by: Dmitry Savitskiy <dmitry.savitskiy@datacore.com>